chore: Extract `IGrid` into its own header
